refactor: 잘못 명명한 Model -> CarService로 수정
