+++ src/main/kotlin/racingcar/Application.kt
@@ -1,11 +1,11 @@
 package racingcar
 
 import racingcar.controller.CarController
+import racingcar.service.CarService
 import racingcar.view.InputView
+import racingcar.view.OutputView
 
 fun main() {
-    val view = InputView()
-    val model = Model()
-    val controller = CarController(model, view)
+    val controller = CarController(InputView(), OutputView())
     controller.run()
 }
